(feature) Add AssertRules for Housekeeping Tasks
The conditions are:

If the given department is ‘Housekeeping’, the job item is one of ['Blanket', 'Sheets', 'Mattress'], and a location with a location type of 'Room' is specified, then create a job to clean the bed(s) in the given room.
If a location with a location type of 'Floor' is specified, then create a job to clean the beds in all rooms of type ‘Room’ on that floor.
If the given location is neither of location type ‘Room’ nor ‘Floor’, then return a bad request."