(feature) Add AssertRules for Engineering Tasks
The conditions are:

If the department is ‘Engineering’, a job_item is provided, and at least one location is specified, then create a job to repair the given job item at the specified location(s).
If only one location is given and it is of location type ‘Floor’, then create a job to repair the given job item in all locations on that floor.