(feature) Add AssertRules for 'Room Service' Tasks
The conditions are:

If the department is ‘Room Service’, a job item is provided, and at least one location is specified, then create a job to deliver that job item to the given locations.
If only one location is given and it is of location type ‘Floor’, then create a job to deliver the job item to all locations with a location type of 'Room' on that floor.